wip on implementation

+++ core/src/me/brucefreedy/adofai/ParseUnit.kt
@@ -4,7 +4,7 @@ import com.google.gson.JsonObject
 import me.brucefreedy.game.Angle
 
 data class ParseUnit(
-    var bpm: Int,
+    var bpm: Double,
     var floor: Int = 0,
     val angleList: List<Angle>,
     var json: JsonObject,
